Renaming random_engine.* to random.*
