+++ include/arrayfire.h
@@ -312,7 +312,7 @@
 #include "af/image.h"
 #include "af/index.h"
 #include "af/lapack.h"
-#include "af/random_engine.h"
+#include "af/random.h"
 #include "af/seq.h"
 #include "af/signal.h"
 #include "af/sparse.h"
