+++ src/api/c/random.cpp
@@ -7,7 +7,7 @@
  * http://arrayfire.com/licenses/BSD-3-Clause
  ********************************************************/
 
-#include <af/random_engine.h>
+#include <af/random.h>
 
 #include <af/array.h>
 #include <af/data.h>
